Fix handling of start index in `str.indexof_re`
Fixes #6636, #6637. When the start index was non-zero, the result of
`str.indexof_re` was not properly restricted to be greater or equal to
the start index. This commit fixes the issue by making the eager
reduction lemma more precise. Additionally, the commit fixes an issue
with the lower bound of the length of the match in `str.indexof_re`.