Remove unwanted side effects in `SPLIT_EQ_STRIP_L`
Fixes #6681. When checking whether `SPLIT_EQ_STRIP_L` applies, we were
using `sripSymbolicLength`, which changes its inputs depending on how
many elements of the concatenation it could strip. However, one of the
arguments, `pfxv0`, was reused across multiple checks if the strip did
not result in a rewrite. Later invocations were wrong as a result. This
commit changes the call to `stripSymbolicLength()` to use a new copy of
the vector instead.