Add getDomainElements to C++ and Python APIs
Add the relevant methods to the Python API (src/api/python/cvc5.
{pxi,pxd}), C++ API (src/api/cpp/cvc5.{h,cpp}), and SMT engine
(src/smt/smt_engine.{h,cpp}).

Add missing checks to getValue(Term) in C++ API, so it is consistent
with getValue(vector<Term>).

Change mkGroundTerm to mkGroundValue in TheoryModel::getDomainElements
so that the result of getDomainElements is consistent with getValue
when an element needs to be synthesized (because e.g. no assertions are
made about that sort).

Signed-off-by: Jason Koenig <jrkoenig@stanford.edu>