Fix initialization of `ConjectureGenerator`
Commit 525fe1f4f33aca05f5e92b0cc0f3e0c7e6effa8b removed a flag in the
equality engine that prevented the `EqualityEngine` from notifying
others about the terms created during initialization. This was a problem
in `ConjectureGenerator` because
`ConjectureGenerator::eqNotifyNewClass()` pushes to `d_upendingAdds`,
which is created after the `EqualityEngine`. There are at least three
ways to solve this:

- Revert commit 525fe1f4f33aca05f5e92b0cc0f3e0c7e6effa8b, which would
  add back additional checks in the `EqualityEngine`.
- Move `d_upendingAdds` before `d_uequalityEngine`, which seems a bit
  brittle, especially because the constructor creates a class that calls
  back into the partially constructed class (AFAIK the C++ standard
  allows it, though).
- Add an `init()` method for `QuantifiersModule`, which allows it to
  initialize additional members that require the full class to be
  initialized.

This commit implements the third option since it seemed relatively
straightforward. The downside is that everyone using an `EqualityEngine`
has to be aware that notifications already happen when it is created.