Configuration: Indicate dependencies being built
This commit changes the output of our configuration to show which
dependencies are being built as part of the cvc5 build. For example:

```
ABC                       : off
CryptoMiniSat             : off
GLPK                      : off
Kissat                    : off
LibPoly                   : on (building)
CoCoALib                  : off
```

Indicates that `LibPoly` was not found in the system and is thus built
as part of the cvc5 build.